% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatcher.R
\name{add_error_handler}
\alias{add_error_handler}
\title{add_error_handler}
\usage{
add_error_handler(callback)
}
\arguments{
\item{callback}{A function that takes \code{(Bot, Update)} as arguments.}
}
\description{
Registers an error handler in the \code{\link{Dispatcher}}.
}
