% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefilter.R
\name{BaseFilter}
\alias{BaseFilter}
\alias{as.BaseFilter}
\alias{is.BaseFilter}
\title{The base of all filters}
\usage{
BaseFilter(filter)

as.BaseFilter(x, ...)

is.BaseFilter(x)
}
\arguments{
\item{filter}{If you want to create your own filters you can call this generator passing by
a \code{filter} function that takes a \code{message} as input and returns a boolean:
\code{TRUE} if the message should be handled, \code{FALSE} otherwise.}
}
\description{
Base class for all Message Filters.
}
\details{
See \code{\link{filtersLogic}} to know more about combining filter functions.
}
\examples{
\dontrun{

}
}
