% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bot_token}
\alias{bot_token}
\title{bot_token}
\usage{
bot_token(bot_name)
}
\arguments{
\item{bot_name}{The bot's name}
}
\description{
Obtain token from system variables (in \code{Renviron}) set
according to the naming convention \code{R_TELEGRAM_BOT_X}
where \code{X} is the bot's name.
}
\examples{
\dontrun{
bot_token('RBot')
}
}
