% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbackqueryhandler.R
\docType{class}
\name{CallbackQueryHandler}
\alias{CallbackQueryHandler}
\title{Handling callback queries}
\format{
An \code{\link{R6Class}} object.
}
\usage{
CallbackQueryHandler(callback, pattern = NULL)
}
\arguments{
\item{callback}{The callback function for this handler.
See \code{\link{Handler}} for information about this function.}

\item{pattern}{(Optional). Regex pattern to test.}
}
\description{
\code{\link{Handler}} class to handle Telegram callback queries. Optionally
based on a regex.
}
