% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{check_update}
\alias{check_update}
\title{Check an update}
\usage{
check_update(update)
}
\arguments{
\item{update}{The update to be tested.}
}
\description{
This method is called to determine if an update should be handled by
this handler instance. It should always be overridden (see
\code{\link{Handler}}).
}
