% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandhandler.R
\docType{class}
\name{CommandHandler}
\alias{CommandHandler}
\title{CommandHandler}
\format{An \code{\link{R6Class}} object.}
\usage{
CommandHandler(command, callback, filters = NULL, pass_args = FALSE)
}
\arguments{
\item{command}{The command or list of commands this handler
should listen for.}

\item{callback}{The callback function for this handler.
See \code{\link{Handler}} for information about this function.}

\item{filters}{(Optional). Only allow updates with these Filters. See
\code{\link{Filters}} for a full list of all available filters.}

\item{pass_args}{(Optional). Determines whether the handler should be passed
\code{args}, received as a \code{vector}, split on spaces.}
}
\description{
\code{\link{Handler}} class to handle Telegram commands.
}
