% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{delete_webhook}
\alias{delete_webhook}
\title{delete_webhook}
\usage{
delete_webhook()
}
\description{
Use this method to remove webhook integration if you decide to switch back to
\code{getUpdates}. Requires no parameters.
}
