% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{deleteWebhook}
\alias{deleteWebhook}
\title{Remove webhook integration}
\usage{
deleteWebhook()
}
\description{
Use this method to remove webhook integration if you decide to switch back
to \code{getUpdates}. Requires no parameters.
}
\details{
You can also use it's snake_case equivalent \code{delete_webhook}.
}
