% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatcher.R
\docType{class}
\name{Dispatcher}
\alias{Dispatcher}
\alias{is.Dispatcher}
\title{The dispatcher of all updates}
\format{An \code{\link{R6Class}} object.}
\usage{
Dispatcher(bot)

is.Dispatcher(x)
}
\arguments{
\item{bot}{The bot object that should be passed to the handlers.}

\item{x}{Object to be tested.}
}
\description{
This class dispatches all kinds of updates to its registered handlers.
}
\section{Methods}{
 \describe{
    \item{\code{\link{add_handler}}}{Registers a handler in the
      \code{Dispatcher}.}
    \item{\code{\link{add_error_handler}}}{Registers an error handler in the
      \code{Dispatcher}.}
}
}

