% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{editMessageReplyMarkup}
\alias{editMessageReplyMarkup}
\title{editMessageReplyMarkup}
\usage{
editMessageReplyMarkup(chat_id = NULL, message_id = NULL,
  inline_message_id = NULL, reply_markup = NULL)
}
\arguments{
\item{chat_id}{(Optional). Unique identifier for the target chat or username
of the target channel}

\item{message_id}{(Optional). Required if inline_message_id is not specified.
Identifier of the sent message}

\item{inline_message_id}{(Optional). Required if chat_id and message_id are not
specified. Identifier of the inline message}

\item{reply_markup}{(Optional). A Reply Markup parameter object, it can be either:
\itemize{
 \item{\code{\link{ReplyKeyboardMarkup}}}
 \item{\code{\link{InlineKeyboardMarkup}}}
 \item{\code{\link{ReplyKeyboardRemove}}}
 \item{\code{\link{ForceReply}}}}}
}
\description{
Use this method to edit only the reply markup of messages sent by the bot or via the bot
(for inline bots).
}
\details{
You can also use it's snake_case equivalent \code{edit_message_reply_markup}.
}
