% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{editMessageText}
\alias{editMessageText}
\title{Edit a text message}
\usage{
editMessageText(chat_id = NULL, message_id = NULL,
  inline_message_id = NULL, text, parse_mode = NULL,
  disable_web_page_preview = NULL, reply_markup = NULL)
}
\arguments{
\item{chat_id}{(Optional). Unique identifier for the target chat or username
of the target channel.}

\item{message_id}{(Optional). Required if inline_message_id is not
specified. Identifier of the sent message.}

\item{inline_message_id}{(Optional). Required if chat_id and message_id are
not specified. Identifier of the inline message.}

\item{text}{New text of the message.}

\item{parse_mode}{(Optional). Send 'Markdown' or 'HTML', if you want
Telegram apps to show bold, italic, fixed-width text or inline URLs in
your bot's message.}

\item{disable_web_page_preview}{(Optional). Disables link previews for links
in this message.}

\item{reply_markup}{(Optional). A Reply Markup parameter object, it can be
 either:
 \itemize{
  \item{\code{\link{ReplyKeyboardMarkup}}}
  \item{\code{\link{InlineKeyboardMarkup}}}
  \item{\code{\link{ReplyKeyboardRemove}}}
  \item{\code{\link{ForceReply}}}
}}
}
\description{
Use this method to edit text messages.
}
\details{
You can also use it's snake_case equivalent
\code{edit_message_text}.
}
