% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{effective_chat}
\alias{effective_chat}
\title{Get the effective chat}
\usage{
effective_chat()
}
\description{
The chat that this update was sent in, no matter what kind of
update this is. Will be \code{None} for \code{inline_query},
\code{chosen_inline_result}, \code{callback_query} from inline messages,
\code{shipping_query} and \code{pre_checkout_query}.
}
