% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{effective_message}
\alias{effective_message}
\title{Get the effective message}
\usage{
effective_message()
}
\description{
The message included in this update, no matter what kind of
update this is. Will be \code{None} for \code{inline_query},
\code{chosen_inline_result}, \code{callback_query} from inline messages,
\code{shipping_query} and \code{pre_checkout_query}.
}
