% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{effective_user}
\alias{effective_user}
\title{Get the effective user}
\usage{
effective_user()
}
\description{
The user that sent this update, no matter what kind of update this
is. Will be \code{NULL} for \code{channel_post}.
}
