% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basefilter.R
\name{filtersLogic}
\alias{filtersLogic}
\alias{!}
\alias{!.BaseFilter}
\alias{&.BaseFilter}
\alias{|.BaseFilter}
\title{Combining filters}
\usage{
\method{!}{BaseFilter}(f)

\method{&}{BaseFilter}(f, g)

\method{|}{BaseFilter}(f, g)
}
\arguments{
\item{f, g}{Arbitrary \code{\link{BaseFilter}} class functions.}
}
\description{
Creates a function which returns the corresponding logical
    operation between what \code{f} and \code{g} return.
}
\details{
See \code{\link{BaseFilter}} and \code{\link{MessageFilters}} for
    further details.
}
\examples{
not_command <- !MessageFilters$command
text_and_reply <- MessageFilters$text & MessageFilters$reply
audio_or_video <- MessageFilters$audio | MessageFilters$video
}
