% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{forwardMessage}
\alias{forwardMessage}
\title{Forward messages of any kind}
\usage{
forwardMessage(chat_id, from_chat_id, message_id, disable_notification = FALSE)
}
\arguments{
\item{chat_id}{Unique identifier for the target chat or username of
the target channel.}

\item{from_chat_id}{Unique identifier for the chat where the
original message was sent.}

\item{message_id}{Message identifier in the chat specified in from_chat_id.}

\item{disable_notification}{(Optional). Sends the message silently. Users
will receive a notification with no sound.}
}
\description{
Use this method to forward messages of any kind.
}
\details{
You can also use it's snake_case equivalent \code{forward_message}.
}
