% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{get_webhook_info}
\alias{get_webhook_info}
\title{get_webhook_info}
\usage{
get_webhook_info()
}
\description{
Use this method to get current webhook status. Requires no parameters.
}
\details{
If the bot is using \code{getUpdates}, will return an object with the url field empty.
}
