% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{getFile}
\alias{getFile}
\title{getFile}
\usage{
getFile(file_id)
}
\arguments{
\item{file_id}{The file identifier}
}
\description{
Use this method to get basic info about a file and prepare it for downloading. For the
moment, bots can download files of up to 20MB in size. It is guaranteed that the link will be
valid for at least 1 hour. When the link expires, a new one can be requested by
calling get_file again.
}
\details{
You can also use it's snake_case equivalent \code{get_file}.
}
