% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{getMe}
\alias{getMe}
\title{Check your bot's information}
\usage{
getMe()
}
\description{
A simple method for testing your bot's auth token. Requires no parameters.
}
\details{
You can also use it's snake_case equivalent \code{get_me}.
}
