% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{getWebhookInfo}
\alias{getWebhookInfo}
\title{Get current webhook status}
\usage{
getWebhookInfo()
}
\description{
Use this method to get current webhook status. Requires no parameters.
}
\details{
If the bot is using \code{getUpdates}, will return an object with the url
field empty.

You can also use it's snake_case equivalent \code{get_webhook_info}.
}
