% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{handle_update}
\alias{handle_update}
\title{handle_update}
\usage{
handle_update(update, dispatcher)
}
\arguments{
\item{update}{The update to be handled.}

\item{dispatcher}{The dispatcher to collect optional args.}
}
\description{
This method is called if it was determined that an update should indeed
be handled by this instance. It should also be overridden, but in most
cases call \code{self$callback(dispatcher$bot, update)}, possibly along with
optional arguments.
}
