% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{leaveChat}
\alias{leaveChat}
\title{Leave a chat}
\usage{
leaveChat(chat_id)
}
\arguments{
\item{chat_id}{Unique identifier for the target chat or username of
the target channel.}
}
\description{
Use this method for your bot to leave a group, supergroup or channel.
}
\details{
You can also use it's snake_case equivalent \code{leave_chat}.
}
