% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messagehandler.R
\docType{class}
\name{MessageHandler}
\alias{MessageHandler}
\title{MessageHandler}
\format{An \code{\link{R6Class}} object.}
\usage{
MessageHandler(filters, callback)
}
\arguments{
\item{filters}{Only allow updates with these Filters. Use \code{Filters$all}
for no filtering. See \code{\link{Filters}} for a full list of all available filters.}

\item{callback}{The callback function for this handler.
See \code{\link{Handler}} for information about this function.}
}
\description{
\code{\link{Handler}} class to handle Telegram messages. They might contain text, media or status updates.
}
\examples{
\dontrun{
# No filtering
message_handler <- MessageHandler(Filters$all, callback_method)
}
}
