% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{set_method}
\alias{set_method}
\title{set_method}
\usage{
set_method(name, method)
}
\arguments{
\item{name}{Name of the method to be changed. It can be either
\code{'check_update'} or \code{'handle_update'}.}

\item{method}{The method function you are overriding.}
}
\description{
This method allows you to override the \code{\link{check_update}} and
\code{\link{handle_update}} methods.
}
\examples{
# Example of a Handler
callback_method <- function(bot, update){
  chat_id <- update$effective_chat()$id
  bot$sendMessage(chat_id = chat_id, text = 'Hello')
}

hello_handler <- Handler(callback_method)

# Setting the methods
check_update <- function(update){
  TRUE
}
hello_handler$set_method('check_update', check_update)

handle_update <- function(update, dispatcher){
  self$callback(dispatcher$bot, update)
}
hello_handler$set_method('handle_update', handle_update)
}
