% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\name{set_token}
\alias{set_token}
\title{Change your bot's auth token}
\usage{
set_token(token)
}
\arguments{
\item{token}{The bot's token given by the \emph{BotFather}.}
}
\description{
Use this method to change your bot's auth token.
}
