% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updater.R
\name{start_polling}
\alias{start_polling}
\title{start_polling}
\usage{
start_polling(timeout = 10, clean = FALSE, verbose = FALSE)
}
\arguments{
\item{timeout}{(Optional). Passed to \code{\link{getUpdates}}. Default is 10.}

\item{clean}{(Optional). Whether to clean any pending updates on Telegram servers
before actually starting to poll. Default is \code{FALSE}.}

\item{verbose}{(Optional). If \code{TRUE}, prints status of the pollings. Default is \code{FALSE}.}
}
\description{
Starts polling updates from Telegram.
}
