% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webhook.R
\name{stop_server}
\alias{stop_server}
\title{Stop the webhook server.}
\usage{
stop_server()
}
\description{
Stops listening on the webhook. Requires no parameters.
}
\examples{
\dontrun{
# Example of a 'kill' command
kill <- function(bot, update) {
  bot$sendMessage(
    chat_id = update$message$chat_id,
    text = "Bye!"
  )
  # Stop the webhook
  webhook$stop_server()
}

webhook <- Webhook("https://example.com/webhook", "TOKEN") + CommandHandler("start", start)

webhook$start_server()
}
}
