% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\docType{class}
\name{Update}
\alias{Update}
\title{Update}
\format{An \code{\link{R6Class}} object.}
\usage{
Update(data)
}
\arguments{
\item{data}{Data of the update.}
}
\description{
This object represents an incoming \href{https://core.telegram.org/bots/api#update}{Update}.
}
\section{Methods}{
 \describe{
    \item{\code{\link{effective_user}}}{To get the user that sent
    this update, no matter what kind of update this is.}
    \item{\code{\link{effective_chat}}}{To get the chat that this
    update was sent in, no matter what kind of update this is.}
    \item{\code{\link{effective_message}}}{To get the message
    included in this update, no matter what kind of  update this is.}}
}

