% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bot.R
\docType{class}
\name{Bot}
\alias{Bot}
\title{Bot}
\format{An \code{\link{R6Class}} object.}
\usage{
Bot(token, base_url = NULL, base_file_url = NULL)
}
\arguments{
\item{token}{Bot's unique authentication.}

\item{base_url}{Telegram Bot API service URL.}

\item{base_file_url}{Telegram Bot API file URL.}
}
\description{
This object represents a Telegram Bot. It inherits from \code{\link{TGBot}}.
Thus, it has implemented all the API methods from that class. It also features
the \code{\link{get_updates}} method, which allows the use of long polling; and
the \code{\link{set_webhook}}, \code{\link{delete_webhook}} and
\code{\link{get_webhook_info}} methods, which allow to manage webhooks.
}
\details{
To take full advantage of this library take a look at \code{\link{Updater}}.
}
\examples{
\dontrun{
bot <- Bot(token = 'TOKEN')
}
}
