% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bot_token}
\alias{bot_token}
\title{Get a token from environment}
\usage{
bot_token(bot_name)
}
\arguments{
\item{bot_name}{The bot's name.}
}
\description{
Obtain token from system variables (in \code{.Renviron}) set
according to the naming convention \code{R_TELEGRAM_BOT_X}
where \code{X} is the bot's name.
}
\examples{
\dontrun{
# Open the `.Renviron` file
file.edit(path.expand(file.path("~", ".Renviron")))
# Add the line (uncomment and replace <bot-token> by your bot TOKEN):
# R_TELEGRAM_BOT_RTelegramBot=<bot-token>
# Save and restart R

bot_token("RTelegramBot")
}
}
