% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telegram.bot-package.R
\docType{package}
\name{telegram.bot}
\alias{telegram.bot}
\alias{telegram.bot-package}
\title{telegram.bot}
\description{
Features a number of tools to make the development of Telegram bots
with R easy and straightforward, providing an easy-to-use interface
that takes some work off the programmer. It is built on top of the
pure API implementation, being an extension of the \code{\link{telegram}}
package, an R wrapper around the \href{http://core.telegram.org/bots/api}{Telegram Bot API}.
}
\details{
In \href{https://ebeneditos.github.io/telegram.bot/}{this page}
you can learn how to build a Bot quickly with this package.
}
\section{Main Classes}{
 \describe{
    \item{\code{\link{Updater}}}{Package main class. This class,
        which employs the class \code{\link{Dispatcher}}, provides a front-end to
        class \code{\link{Bot}} to the programmer, so they can focus on coding the bot.
        Its purpose is to receive the updates from Telegram and to deliver them to said
        dispatcher.}
    \item{\code{\link{Dispatcher}}}{This class dispatches all kinds of updates to its
        registered handlers.}
    \item{\code{\link{Handler}}}{The base class for all update handlers.}
    \item{\code{\link{Bot}}}{This object represents a Telegram Bot.}
}
}

