% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{class}
\name{TelegramObject}
\alias{TelegramObject}
\alias{is.TelegramObject}
\title{The base of telegram.bot objects}
\format{
An \code{\link{R6Class}} generator object.
}
\usage{
is.TelegramObject(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
Base class for most telegram objects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TelegramObject-clone}{\code{TelegramObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TelegramObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TelegramObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TelegramObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
