% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\docType{class}
\name{TelegramObject}
\alias{TelegramObject}
\alias{is.TelegramObject}
\title{The base of telegram.bot objects}
\format{An \code{\link{R6Class}} generator object.}
\usage{
TelegramObject

is.TelegramObject(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
Base class for most telegram objects.
}
\keyword{datasets}
