% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{user_id}
\alias{user_id}
\title{Get a user from environment}
\usage{
user_id(user_name)
}
\arguments{
\item{user_name}{The user's name.}
}
\description{
Obtain Telegram user id from system variables (in \code{.Renviron}) set
according to the naming convention \code{R_TELEGRAM_USER_X}
where \code{X} is the user's name.
}
\examples{
\dontrun{
# Open the `.Renviron` file
file.edit(path.expand(file.path("~", ".Renviron")))
# Add the line (uncomment and replace <user-id> by your Telegram user ID):
# R_TELEGRAM_USER_Me=<user-id>
# Save and restart R

user_id("Me")
}
}
